requirements.txt
boltons==18.0.1
nltk==3.4
numpy==1.14.3
redlock==1.2.0
pandas==0.24.0
tensorflow-gpu==1.12.0
h5py==2.7.1
gpustat==0.5.0
typing==3.6.4
scikit_learn==0.20.2
wandb==0.6.35
jupyter
xlrd
